% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbookAdjustment.R
\name{adjustWithLogbook}
\alias{adjustWithLogbook}
\title{Adjust landings with logbooks}
\usage{
adjustWithLogbook(
  landings,
  logbooks,
  speciesFAO,
  gearCodes = NULL,
  vesselSize = NULL,
  activities = c("FIS", "REL", "SCR")
)
}
\arguments{
\item{landings}{landings as parsed by \code{\link[stox2.7preprocessing]{readLandings}}}

\item{logbooks}{logbooks as parsed by \code{\link[RstoxData]{readErsFile}}}

\item{speciesFAO}{3-alpha FAO species code, e.g. "COD"}

\item{gearCodes}{Gear codes (NS9400) that should be fetched from logbooks. If null, all available gears are included.}

\item{vesselSize}{smallest vessel size to include from logbooks. If null, all availalbe vessels sizes are included.}

\item{activities}{activity-types to include from logbooks. If null, all available activites are included.}
}
\value{
landings with columns FOlat and FOlon added, and adjusted definitions for columns:
\describe{
 \item{Hovedomr책de_kode}{dominant area for trip OR area of fishing operation}
 \item{Lokasjon_kode}{dominant location (sub area) OR location of fishing operation}
 \item{SisteFangstdato}{date of last catch OR date of fishing operation}
 \item{Rundvekt}{live weight (round weight) listed in sales-note OR weight ascribed to fishing operation}
 \item{Produktvekt}{product weight listed in sales-note OR weight ascribed to fishing operation}
 \item{Bruttovekt}{gross weight of landed fish listed in sales-note OR weight ascribed to fishing operation}
 Definitions depend on whether logbook information was available for the sales-note.
}
}
\description{
Adjust landings with information from logbooks for a specific species.
}
\details{
Logbooks contain catch reports for each fishing operation as defined in legislation for different gears.
 Parameters like catch date and area are defined in landings (sales notes) by a summary value for each trip.
 E.g. the dominant area for the trip (the area with more catch), and the date of the last catch of the trip.
 Applying this function alters the definition of some of these parameters to be resolved to fishing operation
 when information is available from logbook records.
 The information from logbooks are not simply copied over, but the landed weight is redistributed
 to reflect the partitioning of weights reported in logbooks
 In addition positions are added when logbook/VMS entries are available
 (columns FOlat and FOlon are added to landings for latitdue and longitude of fishing operation, respectively).

 In general, the difference in definition between logbooks and landing are less likely to be relevant
 for shorter trips. Trip length potential is limited by vessel size.
 Logbook records was first introduced for trawlers, and later for all vessels larger than or equal to 15 m.
 Selection for certain gears or vessel sizes is supported by the parameters 'gearCodes' and 'vesselSize'

 Logbook records may contain records of activities other than fishing. Some of them may relate to
 activity that does not consitute removal of fish from the sea (such as transfer of fish between vessels).
 Selection of activities to include is supported by the parameter 'activities'. Catch is occationally
 recorded for operations which do not handle fish (such as STE for steaming or SET for setting of gear), these
 are likely misrecorded and should possibly be included. Codes for transport and production (HAU and PRO) may
 indicate processing and transport of the catch weight and should probably be excluded. By default only
 codes for fishing (FIS), harvesting from another ships gear (REL) and research-catches (SCR) are included.

 A trip may be reflected in several sales notes. This function distributes the landed
 weight for each of them on the temporal and spatial spatial variables.
 That is: each sales-note line with a corresponding logbook entry is replaced by
 a set of sale-note lines, one for each fishing operation in the corresponding trip.
 This is done regardless of parameters that may be specific to certain fishing operations,
 such as gear, and it is done without attempt to make a realistic partitioning of
 parameters determined at landing, such as usage.

 trips are deduced by vessel identifiers (radio call-signal) and dates (date of last catch
 in landings, and date of fishing operation in logbooks). Consisteecy of gears reported
 between logboks and landings are not enforced.

 Only codes are changed. Corresponding description columns are not. E.g.:
 Area code (Hovedomr책de_kode) may reflect fishing operation after logbook-adustments,
 while area description (Hovedomr책de_bokm책l) will reflect dominant area for trip.
}
