% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coastalCodAreas-datadoc.R,
%   R/coastalCodAreasAFWG-datadoc.R
\docType{data}
\name{coastalCodAreas}
\alias{coastalCodAreas}
\title{Coastal cod areas}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with area names identified in the column 'polygonName'. See \code{\link[RstoxBase]{StratumPolygon}}.

\code{\link[sp]{SpatialPolygonsDataFrame}} with area names identified in the column 'polygonName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(coastalCodAreas)

data(coastalCodAreasAFWG)
}
\description{
Definition for areas used for catch at age estimation of coastal cod.
The areas overlap with area 00, 03, 04, 05, 06, and 07 defined by the Norwegian Directorate of Fisheries
(see: RstoxFDA::mainareaFdir2018), and are constructed from these by dividing them along the Norwegian
territorial line (12 nm of coast).

Definition for areas used for catch at age estimation of coastal cod in AFWG.
This is an adaptation of \code{\link[stox2.7preprocessing]{coastalCodAreas}}
}
\details{
Polygons are defined in WGS84 coordinates (unprojected).

Area positions for use with applyPosToData in Stox 2.7 can be saved with \code{\link[stox2.7preprocessing]{writePolygonPositions}}
Areas may be plotted for inspection with \code{\link[RstoxFDA]{plotArea}} (see example).

Polygons are defined in WGS84 coordinates (unprojected).

The polygons may be saved as a stox 2.7 stratafile with \code{\link[stox2.7preprocessing]{writeSpDataFrameAsWKT}}
Areas may be plotted for inspection with \code{\link[RstoxFDA]{plotArea}} (see example).
}
\examples{
 # plot coastal cod areas
 data(coastalCodAreas)
 RstoxFDA::plotArea(areaDef=coastalCodAreas, title = "Coastal Cod areas")
 # plot coastal cod areas AFWG
 data(coastalCodAreasAFWG)
 RstoxFDA::plotArea(areaDef=coastalCodAreasAFWG, title = "Coastal Cod areas")
}
\keyword{datasets}
