% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{compareLandings}
\alias{compareLandings}
\title{compare landings}
\usage{
compareLandings(original, modified)
}
\arguments{
\item{original}{sales notes to be treatad as original}

\item{modified}{sales notes to be treated as modifed}
}
\value{
table
}
\description{
Make table comparing two sets of sales notes by gear and area
}
