% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coastalCodArea.R
\name{encodeCostalCodArea}
\alias{encodeCostalCodArea}
\title{encode costal cod areas}
\usage{
encodeCostalCodArea(landings)
}
\arguments{
\item{landings}{landings as parsed by \code{\link[stox2.7preprocessing]{readLandings}}}
}
\value{
landings with codes for coastal cod areas in the colunm 'Hovedområde_kode' where relevant
}
\description{
encode costal cod areas in the column for main area
}
\details{
Coastal cod areas are defined as in \code{\link[stox2.7preprocessing]{coastalCodAreas}}
 They are inserted into the column 'Hovedområde_kode' based on its value and the value for the field
 column 'KystHav_kode' which indicate whether landings is from the coastal or off-shore side of the
 terretorial line (12 nm of coast).

 If positions are available for some landings (see \code{\link[stox2.7preprocessing]{adjustWithLogbook}})
 these will be used for setting the area, rather than the columns 'Hovedområde_kode' and 'KystHav_kode'
}
