% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFWGadjustments.R
\name{processLandingsAllAdjustmentsAFWG}
\alias{processLandingsAllAdjustmentsAFWG}
\title{Prepare AFWG cod}
\usage{
processLandingsAllAdjustmentsAFWG(
  fileName,
  landings,
  logbooks,
  seasonalConversionFactor = T,
  coastalCod = T
)
}
\arguments{
\item{fileName}{filename where adjusted landings in XML format will be saved}

\item{landings}{input: landings in XML format}

\item{logbooks}{logbooks in PSV format.}

\item{seasonalConversionFactor}{if TRUE seasonal conversion factors are applied for gutted fish}

\item{coastalCod}{if TRUE only the Norwegain coast AFWG areas are saved, otherwise all other areas are saved.}
}
\description{
Prepares landings for AFWG cod as agreed on WKBarFar 2021.
}
\details{
This function adjust landings by applying seasonal conversion factors,
 logbook adjustment and recoding of Norwegain coast AFWG areas,
 as on WKBarFar 2021.
 Some control parameters are calculated and messaged.

 The resulting adjusted landings may be used by Stox 2.7. In order to get the necessary
 resource files for using the coastal cod areas as stratas see documentation for
 \code{\link[stox2.7preprocessing]{coastalCodAreas}} and \code{\link[stox2.7preprocessing]{coastalCodAreasAFWG}}

 To apply any of these adjustments indvidually
 consider the functions \code{\link[stox2.7preprocessing]{adjustConversionFactor}},
 \code{\link[stox2.7preprocessing]{adjustWithLogbook}} and
 \code{\link[stox2.7preprocessing]{encodeCostalCodArea}}

 To avoid encoding issues introduced earlier in the data-flow
 four fields with names of 'kommune' is removed.

 Default settings prepares coastal cod.
 File writing is quite slow and make take several hours. Filtering on areas may speed this up a little bit.
 Run with 'coastalCod' TRUE to save data only for the Norwegain coast AFWG areas
 Run with 'coastalCod' FALSE to save data only for all other areas (note that further filtering will be required in Stox to get the AFWG area)
 Running with parameter seasonalConversionFactor=FALSE data without seasonal conversion factors.
}
