% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readLandings}
\alias{readLandings}
\title{Read landings as Stox 2.7 does it
Rename variables to Stox 3.0 conventions
Reformat sistefangstdato and landingsdato to StoX 3.0 convention
Keep flat format}
\usage{
readLandings(landingsXML)
}
\arguments{
\item{filename}{for xml-formatted landings}
}
\description{
Read landings as Stox 2.7 does it
Rename variables to Stox 3.0 conventions
Reformat sistefangstdato and landingsdato to StoX 3.0 convention
Keep flat format
}
