% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{writePolygonPositions}
\alias{writePolygonPositions}
\title{Write polygon positions}
\usage{
writePolygonPositions(shape, output)
}
\arguments{
\item{shape}{\code{\link[RstoxBase]{StratumPolygon}} polygons to save positions for}

\item{output}{filename to save output to}
}
\description{
Write polygon positions for use with ApplyPosToData in Stox 2.7
}
