% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{writeSpDataFrameAsWKT}
\alias{writeSpDataFrameAsWKT}
\title{Writes \code{\link[RstoxBase]{StratumPolygon}} as Stox-WKT files (stratafiles)}
\usage{
writeSpDataFrameAsWKT(shape, output)
}
\arguments{
\item{shape}{\code{\link[RstoxBase]{StratumPolygon}} stratadefinition to convert}

\item{output}{filename to save output to}

\item{namecol}{column in shape that contains strata names}
}
\description{
Writes \code{\link[RstoxBase]{StratumPolygon}} as Stox-WKT files (stratafiles)
}
