% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{writeStox27LandingXML}
\alias{writeStox27LandingXML}
\title{save landings as XML (namespace: "http://www.imr.no/formats/landinger/v2")}
\usage{
writeStox27LandingXML(fileName, landings)
}
\arguments{
\item{fileName}{filename to save to}

\item{landings}{landings as read by readLandings}
}
\description{
save landings as XML (namespace: "http://www.imr.no/formats/landinger/v2")
}
\details{
This is a temprary solution for writing landings XML-files. It needs considerable time.
}
