% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorAdjustment.R
\name{adjustConversionFactor}
\alias{adjustConversionFactor}
\title{Recalculates live weight (round weight) for processed products}
\usage{
adjustConversionFactor(
  landings,
  conversionfactor,
  FAOspecies = "COD",
  vesselRange = c(0, 28),
  months = c(1, 2, 3, 4)
)
}
\arguments{
\item{landings}{sales notes to be adjusted}

\item{conversionfactor}{conversion factor to apply.}

\item{FAOspecies}{specification of the species to apply factor to. FAO 3-alpha species code}

\item{vesselRange}{vector of two numeric values denoting the lower range (inclusive) and the upper range (exclusive) of vessel lengths that factor should be applied for. If NULL the factor is applied to all vessel lengths}

\item{months}{vector of month indecies (1 for january) that the conversion factor should be applied for. If NULL the factor is applied to all months}
}
\description{
Recalculates live weight (round weight) for processed products
}
\details{
Live weight is obtained by applying a scalar conversion factor to weights of partially proceesed product
}
