% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{compareLandings}
\alias{compareLandings}
\title{compare landings}
\usage{
compareLandings(original, modified, includeGear = T)
}
\arguments{
\item{original}{sales notes to be treatad as original}

\item{modified}{sales notes to be treated as modifed}

\item{includeGear}{if True, comparisons are done by gear as well as area}
}
\value{
table
}
\description{
Make table comparing two sets of sales notes by gear and area
}
